+++ app/(DashboardLayout)/editFinancial/page.tsx
@@ -10,7 +10,8 @@ import {
     InputAdornment,
     MenuItem
 } from '@mui/material'
-import BaseCard from '../../components/shared/BaseCard';
+import BaseCard from '../components/shared/BaseCard';
+
 
 // import { createTheme, ThemeProvider, styled } from '@mui/material/styles';
 // const Item = styled(Paper)(({ theme }) => ({
